fix(studio-server): share background removal job runner
